:sparkles: #12 getStaticPropsでAPIルートからではなく関数からLGTMデータを取得するように変更
