+++ src/pages/api/lgtm/images.ts
@@ -1,10 +1,11 @@
 import { NextApiHandler, NextApiRequest, NextApiResponse } from 'next';
 import imageData from '../../../utils/imageData';
+import extractRandomImages from '../../../utils/randomImages';
 
-type Images = { id: number; url: string };
+type Image = { id: number; url: string };
 
 type ImagesResponse = {
-  images?: Images[];
+  images?: Image[];
   error?: {
     code: number;
     message: string;
@@ -13,23 +14,11 @@ type ImagesResponse = {
 
 const imageLength = 9;
 
-const getImages = (arr: Images[], n: number): Images[] => {
-  const copy = [...arr];
-  const ret = [];
-
-  for (let i = n; i > 0; i -= 1) {
-    const rand = Math.floor(Math.random() * (copy.length + 1)) - 1;
-    ret.push(...copy.splice(rand, 1));
-  }
-
-  return ret;
-};
-
 const handler: NextApiHandler = (
   _req: NextApiRequest,
   res: NextApiResponse<ImagesResponse>,
 ): void => {
-  const randomImages = getImages(imageData, imageLength);
+  const randomImages = extractRandomImages(imageData, imageLength);
   const imagesResponse = {
     images: randomImages,
   };
